requirements.txt
fastapi
uvicorn[standard]

Pillow
python-dotenv

diffusers
transformers
accelerate
torch

google-genai

requests
tqdm
